Implemented delete profession confirmation modal

+++ client/pages/Professions/index.js
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import { shape, arrayOf, number } from 'prop-types';
-import { useRouter } from 'next/router';
 import { useAlert } from 'react-alert';
 import Layout from 'shared/components/layout';
 import useErrorHandler from 'shared/hooks/useErrorHandler';
+import useListDataFetch from 'shared/hooks/useListDataFetch';
 import { ERROR_MESSAGES, WARNING_MESSAGES, PAGE_NAMES } from 'shared/constants';
 import ProfessionsAPI from 'pages/Professions/api';
 import { ProfessionType } from 'shared/prop-types';
@@ -11,6 +11,7 @@ import { UserType } from 'common/prop-types/authorization/user';
 import SearchField from 'shared/components/searchField';
 import Pagination from 'shared/components/pagination';
 import Pager from 'shared/components/pager';
+import UAPrompt from 'shared/components/prompt';
 import DataGrid from 'shared/components/dataGrid';
 import Filters from 'shared/components/filters';
 import CreateProfessionDrawer from './components/createProfessionDrawer';
@@ -20,21 +21,25 @@ const ProfessionsPage = (props) => {
   const { user, professions, pages } = props;
 
   const handleError = useErrorHandler();
+  const { refetchQuery } = useListDataFetch();
   const alert = useAlert();
-  const router = useRouter();
+
   const [ isLoading, setIsLoading ] = useState(false);
+  const [ professionsToDelete, setProfessionsToDelete ] = useState([]);
+  const [ isDeleteModalOpen, setIsDeleteModalOpen ] = useState(false);
+
+  const canCreate = user.create('professions');
+  const canDelete = user.delete('professions');
 
   const createProfession = async (payload) => {
+    if (!canCreate) return;
     setIsLoading(true);
 
     try {
       const { status } = await ProfessionsAPI.create(payload);
 
       if (status === 201) {
-        await router.replace({
-          pathname: router.pathname,
-          query: router.query,
-        });
+        await refetchQuery();
       }
 
       if (status === 409) {
@@ -47,17 +52,17 @@ const ProfessionsPage = (props) => {
     }
   };
 
-  const onProfessionDelete = async (records) => {
+  const onProfessionDelete = async () => {
+    if (!canDelete) return;
+
+    setIsDeleteModalOpen(false);
     setIsLoading(true);
 
-    const ids = records.map(id => id._id);
-
     try {
+      const ids = professionsToDelete.map(id => id._id);
+
       await ProfessionsAPI.deleteProfessions(ids);
-      await router.replace({
-        pathname: router.pathname,
-        query: router.query,
-      });
+      await refetchQuery();
     } catch (error) {
       handleError(error, ERROR_MESSAGES.PROFESSIONS_DELETE_PROFESSIONS);
     } finally {
@@ -65,9 +70,24 @@ const ProfessionsPage = (props) => {
     }
   };
 
+  const showModal = (records) => {
+    setIsDeleteModalOpen(true);
+    setProfessionsToDelete(records);
+  };
+
+  const hideModal = () => {
+    setIsDeleteModalOpen(false);
+  };
+
+  const deleteProfessionsModalBody = () => (
+    <ul>
+      { professionsToDelete.map(item => <li key={item.name} className="text-dark">{ item.name }</li>) }
+    </ul>
+  );
+
   const headerConfig = [{
     icon: 'delete',
-    action: onProfessionDelete,
+    action: showModal,
   }];
 
   return (
@@ -78,10 +98,12 @@ const ProfessionsPage = (props) => {
 
           <div><SearchField /></div>
 
-          <CreateProfessionDrawer
-            onApply={createProfession}
-            canCreate={user.create('professions')}
-          />
+          { canCreate && (
+            <CreateProfessionDrawer
+              onApply={createProfession}
+              canCreate={user.create('professions')}
+            />
+          )}
         </Layout.Navbar>
 
         <Layout.Content isLoading={isLoading}>
@@ -98,7 +120,6 @@ const ProfessionsPage = (props) => {
             data={professions}
             columns={dataGridColumns}
             headerConfig={headerConfig}
-            onDelete={onProfessionDelete}
           />
         </Layout.Content>
 
@@ -108,6 +129,14 @@ const ProfessionsPage = (props) => {
           <Pagination pages={pages} />
         </Layout.Footer>
       </Layout>
+
+      <UAPrompt
+        show={isDeleteModalOpen}
+        titleContent="Ви дійсно хочете видалити наступні професії?"
+        bodyContent={deleteProfessionsModalBody}
+        onSubmit={onProfessionDelete}
+        onHide={hideModal}
+      />
     </div>
   );
 };
